term-project/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
psutil==5.9.4
psycopg2-binary==2.9.6
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
